src/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
orjson==3.10.18
sqlalchemy[asyncio]==2.0.40
asyncpg==0.30.0
alembic==1.15.2
pydantic-settings==2.9.1
pydantic[email]==2.11.3
PyJWT==2.9.0
bcrypt==4.1.3
httpx==0.28.1
celery[redis]==5.5.1
redis==5.2.1
psycopg2-binary==2.9.10
slowapi==0.1.9
